7/8 Blosum Pt. 2
Fixed the numpy text file IO issue I had yesterday.

I tackled the problem with a little creativity (setting the data type to be the same as a numpy array filled with zeros). With that IO going smoothly, I could make a pretty Panda dataframe with columns and headers. In addition, I added in an additional get_value function that is not contigent off a panda dataframe, or numpy 2D array that would require a quick letter_compression algorithm on the spot. Instead, I created a python dictionary where the key was a tuple of letters and the value was the numerical value associated in the Blosum Matrix.